+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ## [1.2.1] - 2021-10-13
 
 ### Added
+- New `check_consistency` method for `BeamList` objects.
 - Support for writing out measurement set files.
 - Support for unit tests parallelized with MPI.
 - Require that future changes not drastically increase runtime for current capabilities.